Move test majority networks
